+++ modules/animals/test/animals.mock.js
@@ -1,0 +1,19 @@
+exports.animalData = {
+  "_id": "5c67ff7d59c98e40d3072e06",
+  "animal": {
+    "size": "small",
+    "animalType": "cat",
+    "color": "white"
+  },
+  "contact": {
+    "contactName": "Jack Jackonson",
+    "contactEmail": "jjson@jjjjssso.com",
+    "contactPhone": 111111111,
+  },
+  "coordinates": [
+    40.22,
+    3.87
+  ],
+  "date": "2019-02-16T12:15:34.431Z",
+  "status": "lost"
+}
